Malf Config Fix + Readme update
Updater readme to refer users to the mod's page, as it was out of date. Made it so magazine and ammo malf stats are only changed if malf changes is set to true, big issue no one reported for some reason. Made discard limit config option along with inraid restrictions.