Tweaks and Stats
Buffed .357 damage. Fixed DVL weight, and shifted some it more rearward. Fixed reactor brake not being of muzzle_supp_adapter type. Added fix for Mechanic quests so that they are all completable. It's set up so that changes to the requirements can be made easily in the future. Lowered torques of all AKs