Fixes + bot loot gen overrides
Fixed BigPipe LO1 crashing server due to bad json. Got "checkRequired" working (at least it doesn't crash, not sure if it's doing its job). Rewrote changes to bot loot to behave like it did in 2.x.x versions: meds,  and other loot in the future, can only spawn in the inventory slot in which they are found rather than populating a global array and dumped into the tac vest. Still need to add backpack meds and conditionals depending if the bot is a boss or not, adjusting loot limits, etc.